feat: add signal handling in http-server

+++ lib/net/net.cpp
@@ -1,4 +1,5 @@
 #include "net.h"
+#include <cstdint>
 #include <cstdio>
 
 LIBNET_EXPORT const char *libnet_resolve(const char *domain_name) {
@@ -26,7 +27,7 @@ LIBNET_EXPORT const char *libnet_resolve(const char *domain_name) {
     return ip_str;
 }
 
-LIBNET_EXPORT int libnet_socket(int64_t *result) {
+LIBNET_EXPORT int64_t libnet_socket(int64_t *result) {
     if (int rc = ::socket(AF_INET, SOCK_STREAM, 0); rc < 0) {
         return rc;
     } else {
@@ -35,7 +36,7 @@ LIBNET_EXPORT int libnet_socket(int64_t *result) {
     }
 }
 
-LIBNET_EXPORT int libnet_socket_connect(int fd, const char *ipAddr, uint16_t port) {
+LIBNET_EXPORT int64_t libnet_socket_connect(int fd, const char *ipAddr, uint16_t port) {
     struct sockaddr_in serv_addr{};
     serv_addr.sin_family = AF_INET;
     serv_addr.sin_port = htons(port);
@@ -48,7 +49,7 @@ LIBNET_EXPORT int libnet_socket_connect(int fd, const char *ipAddr, uint16_t por
     return 0;
 }
 
-LIBNET_EXPORT int libnet_socket_bind(int fd, const char *ipAddr, uint16_t port) {
+LIBNET_EXPORT int64_t libnet_socket_bind(int fd, const char *ipAddr, uint16_t port) {
     struct sockaddr_in serv_addr{};
     serv_addr.sin_family = AF_INET;
     serv_addr.sin_port = htons(port);
@@ -61,7 +62,7 @@ LIBNET_EXPORT int libnet_socket_bind(int fd, const char *ipAddr, uint16_t port) 
     return 0;
 }
 
-LIBNET_EXPORT int libnet_socket_accept(int fd, sockaddr_in *clientAddr, socklen_t *clientAddrLen) {
+LIBNET_EXPORT int64_t libnet_socket_accept(int fd, sockaddr_in *clientAddr, socklen_t *clientAddrLen) {
     if (auto nfd = accept(fd, (struct sockaddr *)clientAddr, clientAddrLen); nfd > 0) {
         return nfd;
     } else {
@@ -69,14 +70,14 @@ LIBNET_EXPORT int libnet_socket_accept(int fd, sockaddr_in *clientAddr, socklen_
     }
 }
 
-LIBNET_EXPORT int libnet_socket_send(int fd, char *buf, u_int64_t len, int sendFlag) {
+LIBNET_EXPORT int64_t libnet_socket_send(int fd, char *buf, u_int64_t len, int sendFlag) {
     if (send(fd, buf, len, sendFlag) < 0) {
         return LIBNET_SEND_FAIL;
     }
     return LIBNET_SUCCESS;
 }
 
-LIBNET_EXPORT int libnet_socket_recv(int fd, char *buf, u_int64_t len, int recvFlag) {
+LIBNET_EXPORT int64_t libnet_socket_recv(int fd, char *buf, u_int64_t len, int recvFlag) {
     if (auto res = recv(fd, buf, len, recvFlag); res < 0) {
         return LIBNET_RECV_FAIL;
     } else if (res == 0) {
@@ -125,14 +126,14 @@ LIBNET_EXPORT sslInfo *libnet_ssl_connect(int fd, const char *domain) {
     return r;
 }
 
-LIBNET_EXPORT int libnet_ssl_send(sslInfo *fd, char *data, u_int64_t len) {
+LIBNET_EXPORT int64_t libnet_ssl_send(sslInfo *fd, char *data, u_int64_t len) {
     if (SSL_write(fd->ssl, data, len) <= 0) {
         return LIBNET_SSL_SEND_FAIL;
     }
     return LIBNET_SUCCESS;
 }
 
-LIBNET_EXPORT int libnet_ssl_recv(sslInfo *fd, char *dest, u_int64_t len) {
+LIBNET_EXPORT int64_t libnet_ssl_recv(sslInfo *fd, char *dest, u_int64_t len) {
     if (auto res = SSL_read(fd->ssl, dest, len); res < 0) {
         return LIBNET_SSL_RECV_FAIL;
     } else if (res == 0) {
@@ -149,7 +150,7 @@ LIBNET_EXPORT void libnet_ssl_close(sslInfo *fd) {
     close_socket(fd->fd);
 }
 
-LIBNET_EXPORT int libnet_socket_listen(int fd, int backlog) {
+LIBNET_EXPORT int64_t libnet_socket_listen(int fd, int backlog) {
     if (int rc = ::listen(fd, backlog); rc < 0) {
         return rc;
     } else {
@@ -161,7 +162,7 @@ LIBNET_EXPORT void libnet_free(void *ptr) {
     return free(ptr);
 }
 
-LIBNET_EXPORT int libnet_ssl_get_error() {
+LIBNET_EXPORT int64_t libnet_ssl_get_error() {
     return ERR_get_error();
 }
